# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ireland location-services API",
    "version": "2.1.0",
    "description": "Local Hono/Bun + PostGIS API. GeoJSON map features and typed source context with explicit provenance. Dataset completeness and rankings are not certified."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3080",
      "description": "Native macOS development containers"
    }
  ],
  "paths": {
...
    "/criteria": {
      "get": {
        "operationId": "criterionRegistry",
        "summary": "Configurable criteria; no fixed selection count",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or snapshot conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Invalid request or snapshot conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/data-status": {
      "get": {
        "operationId": "dataStatus",
        "summary": "Active snapshots, raw adapter status and coverage",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "version": {
                      "const": 2
                    },
                    "totalRecords": {
                      "type": "integer"
                    },
                    "sources": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CommonManifest"
                      }
                    },
                    "adapterRegistry": {
                      "type": "object"
                    },
                    "rankingReady": {
                      "const": false
                    }
                  },
                  "required": [
                    "version",
                    "totalRecords",
                    "sources",
                    "adapterRegistry",
                    "rankingReady"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or snapshot conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Invalid request or snapshot conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/places": {
      "get": {
        "operationId": "findPlaces",
        "summary": "Search indexed CSO urban-area names",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 250,
              "default": 100
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 4096
            },
            "description": "Opaque query-bound cursor. Restart paging after HTTP 409."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "version": {
                      "const": 2
                    },
                    "places": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CommonRecord"
                      }
                    },
                    "nextCursor": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "snapshotIds": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "sources": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CommonManifest"
                      }
                    }
                  },
                  "required": [
                    "version",
                    "places",
                    "nextCursor",
                    "snapshotIds",
                    "sources"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or snapshot conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Invalid request or snapshot conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/market-context": {
      "get": {
        "operationId": "marketObservations",
        "summary": "Exact-geography historical buying/rental statistics",
        "parameters": [
          {
            "name": "geographyCode",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Original publisher geography code, without inferred town joins."
          },
          {
            "name": "geographyCodeSystem",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenure",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "buy",
                "rent"
              ]
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "propertyType",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "house",
                "apartment",
                "all",
                "unknown"
              ]
            }
          },
          {
            "name": "bedroomClass",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "bedrooms",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 250,
              "default": 100
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 4096
            },
            "description": "Opaque query-bound cursor. Restart paging after HTTP 409."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "version": {
                      "const": 2
                    },
                    "records": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CommonRecord"
                      }
                    },
                    "nextCursor": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "snapshotIds": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "sources": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CommonManifest"
                      }
                    }
                  },
                  "required": [
                    "version",
                    "records",
                    "nextCursor",
                    "snapshotIds",
                    "sources"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or snapshot conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Invalid request or snapshot conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/layers": {
      "post": {
        "operationId": "preferenceLayers",
        "summary": "Generate selected GeoJSON and information layers",
        "description": "Any number of distinct registered criteria. Exact marketGeography/contextGeography codes are independent from the pin. Display mappings are in config/layer-categories.json.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LayerPreferences"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "version": {
                      "const": 2
                    },
                    "layers": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "rankingReady": {
                      "const": false
                    }
                  },
                  "required": [
                    "version",
                    "layers",
                    "rankingReady"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or snapshot conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Invalid request or snapshot conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/assess": {
      "post": {
        "operationId": "assessEvidence",
        "summary": "Recorded proximity and explicit unknown evidence",
        "description": "All scores are null while thresholds and source coverage are unapproved. No travel times, household connections, clinical suitability or property availability are inferred.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Preferences"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "version": {
                      "const": 2
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "verticals": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "aggregate": {
                      "type": "object"
                    }
                  },
                  "required": [
                    "version",
                    "results",
                    "verticals",
                    "aggregate"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or snapshot conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Invalid request or snapshot conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CommonManifest": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "const": 2
          },
          "snapshotId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "sourceId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "scope": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "publisher": {
            "type": "string",
            "minLength": 1
          },
          "sourceUrl": {
            "type": "string",
            "format": "uri"
          },
          "license": {
            "type": "string",
            "enum": [
              "CC-BY-4.0",
              "CC0-1.0",
              "CC-BY-SA-4.0",
              "ODbL-1.0",
              "public-domain",
              "custom-reuse",
              "unverified"
            ]
          },
          "licenseEvidenceUrl": {
            "type": "string",
            "format": "uri"
          },
          "attribution": {
            "type": "string",
            "minLength": 1
          },
          "reuseStatus": {
            "type": "string",
            "enum": [
              "approved",
              "unverified",
              "restricted"
            ]
          },
          "fetchedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
          },
          "observedPeriod": {
            "type": [
              "string",
              "null"
            ]
          },
          "geography": {
            "type": "string",
            "minLength": 1
          },
          "coverage": {
            "type": "string",
            "enum": [
              "partial",
              "unknown"
            ]
          },
          "limitations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "adapter": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "minLength": 1,
                "maxLength": 1000
              },
              "version": {
                "type": "string",
                "minLength": 1,
                "maxLength": 1000
              }
            },
            "required": [
              "id",
              "version"
            ],
            "additionalProperties": false
          },
          "rawAssets": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string",
                  "minLength": 1
                },
                "sha256": {
                  "type": "string",
                  "pattern": "^[a-f0-9]{64}$"
                },
                "bytes": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                }
              },
              "required": [
                "path",
                "sha256",
                "bytes"
              ],
              "additionalProperties": false
            }
          },
          "recordCount": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          },
          "format": {
            "type": "string",
            "const": "ndjson"
          },
          "quality": {
            "type": "object",
            "properties": {
              "inputRecords": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "acceptedRecords": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "quarantinedRecords": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "unlocatedRecords": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "excludedRecords": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9007199254740991
              },
              "exclusions": {
                "type": "object",
                "propertyNames": {
                  "type": "string"
                },
                "additionalProperties": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                }
              }
            },
            "required": [
              "inputRecords",
              "acceptedRecords",
              "quarantinedRecords",
              "unlocatedRecords",
              "excludedRecords",
              "exclusions"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "version",
          "snapshotId",
          "sourceId",
          "scope",
          "publisher",
          "sourceUrl",
          "license",
          "licenseEvidenceUrl",
          "attribution",
          "reuseStatus",
          "fetchedAt",
          "observedPeriod",
          "geography",
          "coverage",
          "limitations",
          "adapter",
          "rawAssets",
          "recordCount",
          "sha256",
          "format",
          "quality"
        ],
        "additionalProperties": false
      },
...
      "LayerPreferences": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "version": {
            "type": "number",
            "const": 1
          },
          "location": {
            "type": "object",
            "properties": {
              "longitude": {
                "type": "number",
                "minimum": -11,
                "maximum": -5
              },
              "latitude": {
                "type": "number",
                "minimum": 51,
                "maximum": 56
              },
              "placeId": {
                "type": "string"
              }
            },
            "required": [
              "longitude",
              "latitude"
            ],
            "additionalProperties": false
          },
          "criteria": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "minLength": 1
                },
                "importance": {
                  "type": "string",
                  "enum": [
                    "required",
                    "preferred"
                  ]
                },
                "weight": {
                  "type": "number",
                  "exclusiveMinimum": 0,
                  "maximum": 100
                },
                "parameters": {
                  "default": {},
                  "type": "object",
                  "properties": {
                    "maximumMinutes": {
                      "type": "number",
                      "exclusiveMinimum": 0
                    },
                    "mode": {
                      "type": "string",
                      "enum": [
                        "driving",
                        "walking",
                        "cycling",
                        "public_transport"
                      ]
                    },
                    "maximumDistanceM": {
                      "type": "number",
                      "exclusiveMinimum": 0
                    },
                    "childAge": {
                      "type": "number",
                      "minimum": 0,
                      "maximum": 18
                    },
                    "facilityId": {
                      "type": "string"
                    },
                    "requiredService": {
                      "type": "string"
                    },
                    "maximumAmountEur": {
                      "type": "number",
                      "exclusiveMinimum": 0
                    },
                    "bedrooms": {
                      "type": "integer",
                      "exclusiveMinimum": 0,
                      "maximum": 9007199254740991
                    },
                    "minimumDownloadMbps": {
                      "type": "number",
                      "exclusiveMinimum": 0
                    }
                  },
                  "additionalProperties": false
                }
              },
              "required": [
                "id",
                "importance"
              ],
              "additionalProperties": false
            }
          },
          "radiusKm": {
            "default": 10,
            "type": "number",
            "exclusiveMinimum": 0,
            "maximum": 50
          },
          "bbox": {
            "$schema": "https://json-schema.org/draft/2020-12/schema",
            "type": "array",
            "prefixItems": [
              {
                "type": "number",
                "minimum": -11,
                "maximum": -5
              },
              {
                "type": "number",
                "minimum": 51,
                "maximum": 56
              },
              {
                "type": "number",
                "minimum": -11,
                "maximum": -5
              },
              {
                "type": "number",
                "minimum": 51,
                "maximum": 56
              }
            ],
            "items": false,
            "minItems": 4,
            "maxItems": 4
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 250,
            "default": 100
          },
          "marketGeography": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "codeSystem": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "codeSystem"
            ]
          },
          "contextGeography": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "codeSystem": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "codeSystem"
            ]
          }
        },
        "required": [
          "version",
          "location",
          "criteria"
        ],
        "additionalProperties": false
      }
    }
  },
...
}